refactor(standards): add standards and update the code
This commit is adding JS standards and editor config and refactoring scripts from bin folder and siteInfo.js to match those standards. It's also adding a check for standards when you run 'npm start' and badge to readme. This is not a best way to check if standards are ok, but it's a beginning.